+++ pyriemann/utils/covariance.py
@@ -1,6 +1,7 @@
 import warnings
 
 import numpy as np
+from scipy.linalg import block_diag
 
 from sklearn.covariance import oas, ledoit_wolf, fast_mcd, empirical_covariance
 from .test import is_square
@@ -230,6 +231,48 @@ def covariances_X(X, estimator='scm', alpha=0.2):
     return covmats / (2 * alpha)  # Eq(10)
 
 
+def block_covariances(X, blocks, estimator='cov'):
+    """Compute block diagonal covariance.
+
+    Calculates block diagonal matrices where each block is a covariance
+    matrix of a subset of channels.
+    Block sizes are passed as a list of integers and can vary. The sum
+    of block sizes must equal the number of channels in X.
+
+    Parameters
+    ----------
+    X : ndarray, shape (n_matrices, n_channels, n_times)
+        Multi-channel time-series.
+    blocks: list of int
+        List of block sizes.
+    estimator : {'cov', 'scm', 'lwf', 'oas', 'mcd', 'sch', 'corr'} \
+        (default: 'scm')
+        Covariance matrix estimator, see
+            :func:`pyriemann.utils.covariance.covariances`.
+
+    Returns
+    -------
+    C : ndarray, shape (n_matrices, n_channels, n_channels)
+        Block diagonal covariance matrices.
+    """
+    est = _check_est(estimator)
+    n_matrices, n_channels, n_times = X.shape
+
+    if np.sum(blocks) != n_channels:
+        raise ValueError('Sum of individual block sizes '
+                         'must match number of channels of X.')
+
+    covmats = np.empty((n_matrices, n_channels, n_channels))
+    for i in range(n_matrices):
+        blockcov, idx_start = [], 0
+        for j in blocks:
+            blockcov.append(est(X[i, idx_start:idx_start+j, :]))
+            idx_start += j
+        covmats[i] = block_diag(*tuple(blockcov))
+
+    return covmats
+
+
 def eegtocov(sig, window=128, overlapp=0.5, padding=True, estimator='cov'):
     """Convert EEG signal to covariance using sliding window"""
     est = _check_est(estimator)
@@ -448,7 +491,7 @@ def coherence(X, window=128, overlap=0.75, fmin=None, fmax=None, fs=None,
 
 
 def normalize(X, norm):
-    """Normalize a set of square matrices, using trace or determinant.
+    """Normalize a set of square matrices, using corr, trace or determinant.
 
     Parameters
     ----------
@@ -456,8 +499,13 @@ def normalize(X, norm):
         The set of square matrices, at least 2D ndarray. Matrices must be
         invertible for determinant-normalization.
 
-    norm : {"trace", "determinant"}
-        The type of normalization.
+    norm : {"corr", "trace", "determinant"}
+        The type of normalization:
+
+        * 'corr': normalized matrices are correlation matrices, with values in
+          [-1, 1] and diagonal values equal to 1;
+        * 'trace': trace of normalized matrices is 1;
+        * 'determinant': determinant of normalized matrices is +/- 1.
 
     Returns
     -------
@@ -467,7 +515,10 @@ def normalize(X, norm):
     if not is_square(X):
         raise ValueError('Matrices must be square')
 
-    if norm == "trace":
+    if norm == "corr":
+        stddev = np.sqrt(np.abs(np.diagonal(X, axis1=-2, axis2=-1)))
+        denom = np.expand_dims(stddev, axis=-2) * stddev[..., np.newaxis]
+    elif norm == "trace":
         denom = np.trace(X, axis1=-2, axis2=-1)
     elif norm == "determinant":
         denom = np.abs(np.linalg.det(X)) ** (1 / X.shape[-1])
@@ -477,6 +528,10 @@ def normalize(X, norm):
     while denom.ndim != X.ndim:
         denom = denom[..., np.newaxis]
     Xn = X / denom
+
+    if norm == "corr":
+        np.clip(Xn, -1, 1, out=Xn)
+
     return Xn