+++ doc/whatsnew.rst
@@ -34,6 +34,12 @@ v0.2.8.dev
 
 - Add a special form covariance matrix :func:`pyriemann.utils.covariance.covariances_X`
 
+- Add masked and NaN means with Riemannian metric: :func:`pyriemann.utils.mean.maskedmean_riemann` and :func:`pyriemann.utils.mean.nanmean_riemann`
+
+- Add ``corr`` option in :func:`pyriemann.utils.covariance.normalize`, to normalize covariance into correlation matrices
+
+- Add block covariance matrix: :class:`pyriemann.estimation.BlockCovariances` and :func:`pyriemann.utils.covariance.block_covariances`
+
 v0.2.7 (June 2021)
 ------------------